requirements.txt
alembic==1.16.4
attrs==25.3.0
aurora-data-api==0.5.0
aws-cdk-lib==2.211.0
aws-cdk.asset-awscli-v1==2.2.242
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.aws-lambda-python-alpha==2.211.0a0
aws-cdk.cloud-assembly-schema==48.4.0
boto3==1.40.8
botocore==1.40.8
cattrs==25.1.1
certifi==2025.8.3
charset-normalizer==3.4.3
constructs==10.4.2
dotenv==0.9.9
idna==3.10
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.113.0
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
publication==0.0.3
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
SQLAlchemy==2.0.43
sqlalchemy-aurora-data-api==0.5.0
typeguard==2.13.3
typing_extensions==4.14.1
urllib3==2.5.0
